Registration and login system implemented
Registration and login system should now be implemented. User requires
an internet connection to register as it pushes their account to my
server (can change this later). Upon registration it created a local
version of the profile on the device. When logging in, the application
first attempts to login online, receiving either an ok, failed due to
username, or failed in general response. If failed in general response
occurs, the application attempts to log in with a local profile. This
means the user can still use the app when they are not in an area with
internet etc.
Still to do >>
Hash passwords (currently sent as plaintext)
Add user login check to all activities (currently only main)
Code cleanup