few GUI updates and other code updates
MainActivity should be ignored. We are not going to use it anymore.
LoginActivity -> TabbedActivity; code from MainActivity moved to Tabbed
Activity
Updated login screen, tabs for profile / exercises and diets, code for
quiz(still TBD)
the login screen should be a rough guide for modifying the existent
login system(code-wise)